% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avoidable_hospitalisations}
\alias{avoidable_hospitalisations}
\title{Avoidable hospitalisations}
\format{
A data frame with 3600 rows and 5 variables:
\itemize{
\item {municipality} Three-letter abbreviation of municipality.
\item {diagnosis} Diagnosis that caused the admission.
\item {month} First day of admission month.
\item {admissions} Number of admissions.
\item {population} Number of people in target population.
}
}
\usage{
avoidable_hospitalisations
}
\description{
Potentially avoidable hospitalisations in 65+ years citizens from five
diagnoses and nine municipalities.
}
\keyword{datasets}
