% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacteremia_mortality}
\alias{bacteremia_mortality}
\title{Bacteremia Mortality}
\format{
A data frame with 143 rows and 4 columns:
\describe{
\item{hospital}{Hospital code.}
\item{month}{First day of month}
\item{deaths}{Number of bacteremia patients who died within 30 days from infection.}
\item{cases}{Number of patients with bacteremia.}
}
}
\usage{
bacteremia_mortality
}
\description{
30-day mortality after bacteremia.
}
\keyword{datasets}
