% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csection}
\alias{csection}
\title{Ceasearian Section Delay}
\format{
\subsection{\code{csection}}{

A data frame with 24 rows and 3 columns:
\describe{
\item{month}{First day of month.}
\item{avg_delay}{Time in minutes from decision to delivery.}
\item{n}{Number of C-sections.}
}
}
}
\usage{
csection
}
\description{
Delay times for grade 2 C-sections
}
\keyword{datasets}
