% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hba1c}
\alias{hba1c}
\title{Diabetes HbA1c}
\format{
A data frame with 43 columns and 3 variables:
\describe{
\item{month}{month of measurements.}
\item{avg_hba1c}{average of HbA1c measurements (mmol/mol).}
\item{n}{number of patients who visited the clinic}
}
}
\usage{
hba1c
}
\description{
HbA1c measurements in children with diabetes
}
\keyword{datasets}
