% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospital_mortality}
\alias{hospital_mortality}
\title{Hospital Mortality}
\format{
\subsection{\code{hospital_mortality}}{

A data frame with 135 rows and 4 columns:
\describe{
\item{hospital}{Hospital.}
\item{month}{First day of month.}
\item{deaths}{Number of patients who died during hospitalisation.}
\item{discharges}{Number of discharged patients.}
}
}
}
\usage{
hospital_mortality
}
\description{
Patients who die during hospitalisation
}
\keyword{datasets}
