% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ontime_ct}
\alias{ontime_ct}
\title{On-time CT scan}
\format{
A data frame with 24 rows and 3 columns:
\describe{
\item{month}{First day of month.}
\item{ontime}{Number of patients scanned within 3 hours.}
\item{cases}{Number of patients with acute abdomen.}
}
}
\usage{
ontime_ct
}
\description{
Patients with acute abdomen CT scanned within 3 hours after arrival.
}
\keyword{datasets}
