% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbc.R
\name{pbc}
\alias{pbc}
\title{Process behaviour chart}
\usage{
pbc(
  x,
  num = NULL,
  den = 1,
  facet = NULL,
  data = NULL,
  chart = c("run", "i", "ms"),
  freeze = NULL,
  split = NULL,
  exclude = NULL,
  multiply = 1,
  ncol = NULL,
  ylim = NULL,
  yfixed = TRUE,
  title = "",
  xlab = "Subgroup",
  ylab = "Value",
  plot = TRUE
)
}
\arguments{
\item{x}{Numeric or date(time) vector of subgroup values to plot along the x
axis. Or, if y is NULL, x values will be used for y coordinates.}

\item{num}{Numeric vector of measures or counts to
plot on the y axis (numerator).}

\item{den}{Numeric vector of subgroup sizes (denominator).}

\item{facet}{Character (or factor) vector used for faceting.}

\item{data}{Data frame containing the variables used in the plot.}

\item{chart}{Character value indicating the chart type. Possible values are:
'run' (default), 'i', and 'ms'.}

\item{freeze}{Integer indicating the index of the last subgroup in the
baseline period (phase 1). Ignored if split is not NULL.}

\item{split}{Integer indicating the index of the last subgroup before
splitting the graph.}

\item{exclude}{Integer vector of indices to exclude from calculations.}

\item{multiply}{Number to multiply y axis by, e.g. 100 to get percentages
rather than proportions.}

\item{ncol}{Number of columns in faceted plot.}

\item{ylim}{Numeric vector (y1, y2) setting the y axis limits. Useful e.g.
for preventing negative negative control limits (c(0, NA) or
proportions above 1 (100\%) (c(0, 1).}

\item{yfixed}{Logical, if TRUE (default) makes a common scale for y
axes.}

\item{title, xlab, ylab}{Characters setting the main chart title and axis
labels. Use NULL to suppress.}

\item{plot}{Logical, if TRUE (default), plots an SPC chart.}
}
\value{
Invisibly returns a list of class 'pbc'.
}
\description{
Draws run and control charts.
}
\details{
See examples at \url{https://github.com/anhoej/pbcharts/}.
}
\examples{
# Plot a run chart from 12 random normal values
pbc(rnorm(12))

# Plot a control chart from 12 random normal values
pbc(rnorm(12), chart = 'i')

}
