% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbcharts-package.R
\docType{package}
\name{pbcharts-package}
\alias{pbcharts}
\alias{pbcharts-package}
\title{pbcharts: Process Behaviour Charts}
\description{
Run charts and individuals control charts for statistical quality control and improvement. Control limits accommodate varying subgroup sizes as suggested by Taylor (2017) \url{https://variation.com/normalized-individuals-control-chart/} making `pbcharts` useful for a wide range of measurement and count data and a convenient replacement for classic Shewhart control charts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anhoej/pbcharts}
}

}
\author{
\strong{Maintainer}: Jacob Anhøj \email{jacob@anhoej.net}

}
\keyword{internal}
