% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pbc.R
\name{summary.pbc}
\alias{summary.pbc}
\title{Summarise pbc object}
\usage{
\method{summary}{pbc}(object, ...)
}
\arguments{
\item{object}{A 'pbc' object.}

\item{...}{For compatibility with generic summary function.}
}
\description{
Summarise pbc object
}
