% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{waiting_times}
\alias{waiting_times}
\title{Waiting times}
\format{
A dataframe with 630 rows and 5 variables.
\describe{
\item{quarter}{First day of quarter.}
\item{region}{Region.}
\item{operation}{Type of operation.}
\item{avg_days}{Average waiting time in days.}
\item{n}{Number of operations.}
}
}
\usage{
waiting_times
}
\description{
Average waiting times for 5 types of elective surgery in 5 regions.
}
\keyword{datasets}
