% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbc.R
\name{pbc}
\alias{pbc}
\title{Process behaviour chart}
\usage{
pbc(
  x,
  num = NULL,
  den = 1,
  facet = NULL,
  data = NULL,
  chart = c("run", "i"),
  multiply = 1,
  ylim = NULL,
  ncol = NULL,
  freeze = NULL,
  title = NULL,
  xlab = "Subgroup",
  ylab = "Value",
  partlabs = c("phase 1", "phase 2"),
  plot = TRUE,
  print = FALSE
)
}
\arguments{
\item{x}{Numeric or date(time) vector of subgroup values to plot along the x
axis. Or, if y is NULL, x values will be used for y coordinates.}

\item{num}{Numeric vector of measures or counts to
plot on the y axis (numerator).}

\item{den}{Numeric vector of subgroup sizes (denominator).}

\item{facet}{Character (or factor) vector used for faceting.}

\item{data}{Data frame containing the variables used in the plot.}

\item{chart}{Character value indicating the chart type. Possible values are:
'run' (default), and 'i'.}

\item{multiply}{Number to multiply y axis by, e.g. 100 to get percentages rather
than proportions.}

\item{ylim}{Numeric vector (y1, y2) setting the y axis limits. Useful e.g.
for preventing negative negative control limits (c(0, NA) or
proportions above 1 (100\%) (c(0, 1).}

\item{ncol}{Number of columns in faceted plot.}

\item{freeze}{Integer indicating the index of the last subgroup in the
baseline period (phase 1).}

\item{title}{Character setting the main chart title}

\item{xlab}{Character setting the main y axis label}

\item{ylab}{Character setting the main x axis label}

\item{partlabs}{Character vector of length two setting the labels for phase 1
and phase 2 periods.}

\item{plot}{Logical, if TRUE (default), plots an SPC chart.}

\item{print}{Logical, if TRUE, prints a data frame with coordinates.}
}
\description{
Process behaviour chart
}
