% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pbc.R
\name{plot.pbc}
\alias{plot.pbc}
\title{Plot process behaviour chart}
\usage{
\method{plot}{pbc}(x, ...)
}
\arguments{
\item{x}{List of class 'pbc'.}

\item{...}{Unused, necessary for generic function.}
}
\value{
NULL, plots a process behaviour chart.
}
\description{
Plot process behaviour chart
}
\examples{
p <- pbc(rnorm(12), plot = FALSE)
plot(p)
}
